Update airqalibrate.md
I've encountered a couple of spelling errors on the https://wiki.airqo.net/#/calibration/airqalibrate page of the website. The word 'show' is currently misspelled, and it should be corrected to 'should.' Additionally, the word 'will' is also misspelled and should be corrected to 'with.' These minor typos need to be addressed to enhance website quality and user experience.