feat: Add support for fetching grids summary details
This commit adds a new function `getGirdsSummaryDetails` to the `DeviceRegistry.js` file in the `core/apis` directory. This function makes an API call to retrieve the summary details of grids. It uses the `GRIDS_SUMMARY_URL` constant from the `urls/deviceRegistry` module.

Additionally, the `index.jsx` file in the `pages/map` directory has been updated to dispatch the `getGridsDataSummary` action from the `GridsSlice` module. The retrieved grids data is then used to populate the `siteDetails` state.

This change addresses the need to fetch and display summary details of grids in the map page.