chore: Update webpack configuration for website
The webpack.dev.config.js and webpack.fullapp.config.js files were modified to update the webpack configuration for the website. The changes include overriding the public path, setting the static public path to match what Django expects, and updating the entrypoint. These changes improve the development and build processes for the website.