chore: Update TopBarSearch component with debounce functionality
Add debounce functionality to the handleSearch function in the TopBarSearch component. This improves the search experience by delaying the search request until the user has finished typing, reducing unnecessary API calls. Also, update the propTypes to require the onSearch and onClearSearch functions, and add a debounceTime prop to control the debounce delay. Finally, refactor the component to use React.memo instead of the deprecated memo function.